Conflicts:
	app/display/editor/src/main/java/org/csstudio/display/builder/editor/app/NewDisplayContextMenuEntry.java
	app/filebrowser/src/main/java/org/phoebus/applications/filebrowser/FileBrowserController.java

+++ app/save-and-restore/ui/src/main/java/org/phoebus/applications/saveandrestore/ui/model/Threshold.java
@@ -56,7 +56,7 @@ public class Threshold<T extends Number> implements Serializable {
         } else if (n instanceof Float) {
             return (U) Float.valueOf(-n.floatValue());
         } else if (n instanceof Double) {
-            return (U) Double.valueOf(-n.byteValue());
+            return (U) Double.valueOf(-n.doubleValue());
         }
         throw new IllegalArgumentException(String.format("Cannot negate the value %s.", String.valueOf(n)));
     }
